+++ src/GridDomain.Node.Tests/Cat.cs
@@ -41,7 +41,7 @@ namespace GridDomain.Node.Tests
             Version++;
         }
 
-        public int Version { get; private set; }
+        public long Version { get; private set; }
         public Task<IReadOnlyCollection<IDomainEvent>> Execute(ICommand command)
         {
             switch (command)
